+++ releasenotes.md
@@ -6,6 +6,7 @@
 - move video modal to the top of the video detail page
 - comment out the play video on server buttons cuz the stream is prob just better
 - improve tag entry ui with suggestions and single tag entry
+- completely remove the old play video on server
 
 ## 0.4
 - stream video mode